Small cleanups around the LED support
Use `LED_COUNT` instead of `NUM_LEDS`, and move the `leds` array inside the
`Shortcut` object. While there, fix the order of colors in `cRGB` (it should be
grb for the LEDs in the Shortcut), and introduce a CRGB macro, which accepts rgb
args, and creates a `cRGB` instance with the colors in the right order.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>